MyExCode/Web_Django/Django_LineBotTim/requirement.txt
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
future==0.18.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
pycodestyle==2.7.0
requests==2.25.1
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
